added package file. renamed app to index and deleted empty index. renamed seeds.js to seed
