--- app.js
@@ -1,42 +1,0 @@
-const path = require('path');
-const express = require('express');
-const session = require('express-session');
-const SequelizeStore = require('connect-session-sequelize')(session.Store);
-const exphbs = require('express-handlebars');
-const routes = require('./controllers');
-const helpers = require('./utils/helpers');
-const sequelize = require('./config/connection');
-const hbs = exphbs.create({ helpers });
-const app = express();
-
-const PORT = process.env.PORT || 3001;
-
-// Set up sessions with cookies
-const sess = {
-    secret: 'Super secret secret',
-    cookie: {},
-    resave: false,
-    saveUninitialized: true,
-    store: new SequelizeStore({
-        db: sequelize
-    })
-};
-
-app.use(session(sess));
-
-// Set up handlebars.js engine with custom helpers
-app.engine('handlebars', hbs.engine);
-app.set('view engine', 'handlebars');
-
-// Express middleware
-app.use(express.json());
-app.use(express.urlencoded({ extended: true }));
-app.use(express.static(path.join(__dirname, 'public')));
-
-// Turn on routes
-app.use(routes);
-
-// Turn on connection to db and server
-sequelize.sync({ force: false }).then(() => {
-    app.listen(PORT, () => console.log(`Now listening on port ${PORT}!`));
-});
