+++ package.json
@@ -1,0 +1,24 @@
+{
+    "name": "the-recipe-drawer",
+    "version": "1.0.0",
+    "description": "",
+    "main": "index.js",
+    "scripts": {
+      "test": "echo \"Error: no test specified\" && exit 1",
+      "start": "node index.js",
+      "seed": "node seeds/seed.js"
+    },
+    "author": "",
+    "license": "ISC",
+    "dependencies": {
+      "bcrypt": "^5.0.0",
+      "connect-session-sequelize": "^7.0.4",
+      "dotenv": "^8.2.0",
+      "express": "^4.17.1",
+      "express-handlebars": "^5.2.0",
+      "express-session": "^1.17.1",
+      "handlebars": "^4.7.6",
+      "mysql2": "^2.2.5",
+      "sequelize": "^6.3.5"
+    }
+  }
